Fix identifiable scan behaviors and add hybrid scan capability.
