+++ docs/ReleaseHistory.md
@@ -11,6 +11,10 @@
 - FPS => False positive reduction in static analysis.
 - FNS => Flase negative reduction in static analysis.
 
+# 1.4.25 - 06/04/2024
+- BUG: Bring `IdentifiableScan` into precise equivalence with other maskers.
+- NEW: Provide hybrid capability to run high-performance detections in `IdentifiableScan` and fall back to other masker as required.
+
 # 1.4.24 - 06/03/2024
 - RUL: Add `SEC101/060.LooseSasSecret` detection.
 - RUL: Add `SEC101/528.GenericJwt` detection.
