+++ src/Microsoft.Security.Utilities.Core/ISecretMasker.cs
@@ -1,13 +1,14 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
 
 namespace Microsoft.Security.Utilities;
 
 #pragma warning disable SA1600 // Elements should be documented
 
-public interface ISecretMasker
+public interface ISecretMasker : IDisposable
 {
     IEnumerable<Detection> DetectSecrets(string input);
     string MaskSecrets(string input);
