+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/IIdentifiableKey.cs
