+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_101_AzureClientAppLegacyCredential32.cs
