+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_101_AzureClientAppLegacyCredentials34.cs
