+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_109.AzureContainerRegistryLegacyKey.cs
