+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_152_AzureStorageAccountIdentifiableKey.cs
