+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_160_AzureCosmosDbIdentifiableKey.cs
