+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_163_AzureBatchIdentifiableKey.cs
