+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_167_AzureSearchIdentifiableAdminKey.cs
