+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_172_AzureEventHubIdentifiableKey.cs
