+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_176_AzureContainerRegistryIdentifiableKey.cs
