+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_178_AzureIotHubIdentifiableKey.cs
