+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_179_AzureIotDeviceProvisioningIdentifiableKey.cs
