+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_180_AzureIotDeviceIdentifiableKey.cs
