+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_181_AzureApimDirectManaqementKey.cs
