+++ src/Microsoft.Security.Utilities.Core/PreciselyClassifiedSecurityKeys/SEC101_190_AzureEventGridIdentifiableKey.cs
