+++ src/Microsoft.Security.Utilities.Core/UnclassifiedPotentialSecurityKeys/SEC000_002_Unclassified16ByteHexadecimalString.cs
