+++ src/Microsoft.Security.Utilities.Core/UnclassifiedPotentialSecurityKeys/SEC101_060_LooseSasSecret.cs
